BFS_Enerza_Production/BFS_Enerza_Production/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
apscheduler==3.10.4
reportlab==4.2.0
pillow==10.3.0
httpx==0.27.0
redis==5.0.4
celery==5.4.0
boto3==1.34.100
jinja2==3.1.4
python-dotenv==1.0.1
qrcode==7.4.2
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
faker==25.0.1
coverage==7.5.1
